Customization updates and bug fixing
Unsupported attachment types have been removed. The possibility of customizing the work of the bot has also been expanded: the ability to select the time period for posting posts has been added, as well as the maximum number of the best posts that can be taken from each group. A problem with blocked groups was also resolved.